Add server side CSS building and preprocessing

+++ Makefile
@@ -1,5 +1,10 @@
+netlify: build clean
+
+build:
+	npm run build
+
 clean:
-	rm -rfv requirements.txt \
+	[ -z "$CI" ] || rm -rfv requirements.txt \
 		README.md \
 		LICENSE \
 		git.sh \
